% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{buttonColumn}
\alias{buttonColumn}
\title{LCARS button column}
\usage{
buttonColumn(...)
}
\arguments{
\item{...}{\code{lcarsBUtton} objects.}
}
\description{
An LCARS button column is a container for \code{lcarsButton} that can be passed to \code{lcarsBox}. The buttons occur vertically in the left side panel of the box.
}
