% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{lcarsApp}
\alias{lcarsApp}
\title{Launch LCARS demo apps.}
\usage{
lcarsApp(id = "demo")
}
\arguments{
\item{id}{character, app id.}
}
\description{
Currently available apps include: \code{demo}, \code{box}, \code{sweep},
\code{toggle}, \code{elements}.
}
\examples{
\dontrun{lcarsApp("demo")}
}
