% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lcarsdata}
\alias{lcarsdata}
\title{LCARS colors}
\format{A data frame}
\usage{
lcarsdata
}
\description{
A data frame with 32 rows and 3 columns containing color names and values for each of four palette series.
}
\keyword{datasets}
