% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{lcarsHeader}
\alias{lcarsHeader}
\title{LCARS header}
\usage{
lcarsHeader(title = NULL, color = "golden-tanoi",
  title_color = color)
}
\arguments{
\item{title}{character, optional title.}

\item{color}{header color. Any hex color or a named LCARS color.}

\item{title_color}{text color. Any hex color or a named LCARS color.}
}
\description{
A, LCARS header panel.
}
\seealso{
\code{\link{lcarsColors}}.
}
