% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{lcarsWell}
\alias{lcarsWell}
\title{LCARS well}
\usage{
lcarsWell(..., color = "atomic-tangerine",
  background_color = "#000000")
}
\arguments{
\item{...}{panel contents.}

\item{color}{border color. Any hex color or a named LCARS color.}

\item{background_color}{background color. Any hex color or a named LCARS color.}
}
\value{
a div
}
\description{
A simple LCARS well panel wrapper that takes color and background color arguments and understands LCARS color names.
}
