% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_lcars}
\alias{theme_lcars}
\alias{theme_lcars_light}
\alias{theme_lcars_dark}
\title{LCARS ggplot themes}
\usage{
theme_lcars_light(base_size = 11, base_family = "",
  base_line_size = base_size/22, base_rect_size = base_size/22)

theme_lcars_dark(base_size = 11, base_family = "",
  base_line_size = base_size/22, base_rect_size = base_size/22)
}
\arguments{
\item{base_size}{base font size.}

\item{base_family}{base font family.}

\item{base_line_size}{base size for line elements.}

\item{base_rect_size}{base size for rect elements.}
}
\description{
LCARS ggplot themes
}
