% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/button.R
\name{buttonColumn}
\alias{buttonColumn}
\title{LCARS button column}
\usage{
buttonColumn(...)
}
\arguments{
\item{...}{\code{lcarsButton} objects.}
}
\description{
An LCARS button column is a container for \code{lcarsButton} and \code{lcarsToggle} buttons that can be passed to \code{lcarsBox}.
The buttons occur vertically in the left side panel of the box. Toggle button width should be 150px.
}
