% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcars.R
\docType{package}
\name{lcars}
\alias{lcars}
\alias{lcars-package}
\title{lcars: an LCARS R package}
\description{
Library Computer Access/Retrieval System (\href{https://en.wikipedia.org/wiki/LCARS}{LCARS}.
}
\details{
The \code{lcars} package provides simple approximations to 'Library Computer Access/Retrieval System' (LCARS) style and appearance to give plots an LCARS theme.
More details about LCARS can be found at <https://en.wikipedia.org/wiki/LCARS>. The key feature in \code{lcars} is the ability to wrap plots in an LCARS-themed border.
}
