% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcars.R
\docType{package}
\name{lcars}
\alias{lcars}
\title{lcars: LCARS aesthetic for Shiny}
\description{
Library Computer Access/Retrieval System
(\href{https://en.wikipedia.org/wiki/LCARS}{LCARS}.
}
\details{
The \code{lcars} package provides Shiny widgets and Shiny UI theme that
support a 'Library Computer Access/Retrieval System' (LCARS) aesthetic for
Shiny apps. The package also includes functions for adding a minimal LCARS
theme to static \code{ggplot2} graphs.
}
