% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{lcars_rect}
\alias{lcars_rect}
\title{LCARS rectangle}
\usage{
lcars_rect(xmin, xmax, ymin, ymax, color = "atomic-tangerine")
}
\arguments{
\item{xmin}{numeric, left x positions.}

\item{xmax}{numeric, right x positions.}

\item{ymin}{numeric, bottom y positions.}

\item{ymax}{numeric, top y positions.}

\item{color}{fill and border color. Can be any color given in hex format.
Named colors must be LCARS colors. See \code{\link{lcarsdata}} for options.}
}
\value{
draws a rectangle
}
\description{
A simple wrapper around \code{rect}.
}
\examples{
plot(0:1, 0:1)
lcars_rect(0.1, 0.9, 0.6, 0.9)
}
