% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcars.R
\name{lcars_trycatch}
\alias{lcars_trycatch}
\title{Evaluate an expressions, playing appropriate audio samples of LCARS sounds from Star Trek}
\usage{
lcars_trycatch(expr)
}
\arguments{
\item{expr}{An expression to evaluate}
}
\description{
Evaluates and returns the output from an expression. If there expression produces an error or a warning, an appropriate sound effect from Star Trek The Next Generation will be played.
}
\examples{
\dontrun{
# This will produce an error message and sound effect
lcars_trycatch(expr = "Kirk" / "Spock")

# This will produce an output, a warning message, and a sound effect
lcars_trycatch(expr = as.numeric(c("7", "of", "9")))
}

}
