% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box.R
\name{lcarsBox}
\alias{lcarsBox}
\title{LCARS box}
\usage{
lcarsBox(..., title = NULL, subtitle = NULL, buttonColumn = NULL,
  color = "atomic-tangerine", title_color = color,
  subtitle_color = color)
}
\arguments{
\item{...}{box contents.}

\item{title}{character, box title at top right.}

\item{subtitle}{character, box subtitle at bottom right.}

\item{buttonColumn}{optional button column made with \code{buttonColumn}.}

\item{color}{box color. See details.}

\item{title_color}{text title color.}

\item{subtitle_color}{text subtitle color.}
}
\description{
An LCARS box.
}
\details{
Box color can be any color given in hex format. Named colors must be LCARS colors. See \code{\link{lcarsColors}} for options.
Use a vector of three colors for the top, left side, and bottom box panels, respectively.
}
