% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{lcarsPage}
\alias{lcarsPage}
\title{LCARS Shiny UI}
\usage{
lcarsPage(..., title = NULL)
}
\arguments{
\item{...}{The contents of the document body.}

\item{title}{The browser window title (defaults to the host URL of the page).}
}
\value{
A UI definition that can be passed to the shinyUI function.
}
\description{
Create a Shiny UI page with an LCARS theme.
}
